+++ docstring_generator/config.py
@@ -10,12 +10,16 @@ class Config(BaseSettings):
         description='Whether or not to overwrite the existing function docstring',
         default=False,
     )
+    overwrite_class_docstring: Optional[bool] = Field(
+        description='Whether or not to overwrite the existing class docstring',
+        default=False,
+    )
     documentation_style: Optional[str] = Field(
-        description='The format of documentation to use', default='Numpy-Style'
+        description='The format of documentation to use',
+        default='Numpy-Style',
+        enum=['Numpy-Style'],
     )
     directories_ignore: set[str] = Field(
         description='Directories to ignore',
-        default={
-            'venv', '__pycache__', '.git', 'build', 'dist', 'docs'
-        },
+        default={'venv', '__pycache__', '.git', 'build', 'dist', 'docs'},
     )