+++ docstring_generator/helpers.py
@@ -1,12 +1,11 @@
 import ast
 import os
 import subprocess
-from ast import AsyncFunctionDef, Constant, Expr, FunctionDef
-from queue import Queue
-from typing import Optional, Iterator
-from os import path, listdir
-import os
+from ast import AsyncFunctionDef, ClassDef, Constant, Expr, FunctionDef
 from collections import deque
+from os import listdir, path
+from queue import Queue
+from typing import Iterator, Optional
 
 from langchain.prompts import PromptTemplate
 
@@ -45,6 +44,66 @@ def generate_function_docstring(function_code: str) -> str:
     return function_and_docstring
 
 
+class_doc: str = '''
+class MyTestClass:
+    """A class representing a circle with a given radius.
+
+    Attributes:
+        radius (int | float): The radius of the circle.
+
+    Methods:
+        __init__(radius: int | float) -> None:
+            Initializes a MyTestClass object with the given radius.
+
+        calculate_area() -> float:
+            Calculates and returns the area of the circle using the formula: pi * radius^2.
+    """
+
+    def __init__(self, radius: int | float) -> None:
+        """Initializes a MyTestClass object with the given radius.
+
+        Parameters:
+            radius (int | float): The radius of the circle."""
+        self.radius = radius
+
+    def calculate_area(self) -> float:
+        """Calculates and returns the area of the circle.
+
+        Returns:
+            float: The area of the circle using the formula: pi * radius^2."""
+        return pi * self.radius * self.radius
+'''
+
+
+def generate_class_docstring(class_code: str) -> str:
+    class_prompt_template: str = """Write NumPy-style docstrings for the following class and its methods. Do not generate documentation for methods that do not exist: {class_code}"""
+    prompt = PromptTemplate.from_template(template=class_prompt_template)
+    prompt_formatted_str: str = prompt.format(class_code=class_code)
+    class_and_docstring = llm.invoke(prompt_formatted_str)
+    return class_and_docstring
+
+
+def get_class_docstring(class_and_docstring: str) -> str:
+    """Get the class docstring."""
+    class_tree = ast.parse(class_and_docstring)
+    for node in class_tree.body:
+        if isinstance(node, ClassDef):
+            cls_docstring: str = ast.get_docstring(node)
+            return cls_docstring
+
+
+def get_class_methods_docstrings(class_and_docstring: str) -> dict[str, str]:
+    """Get a class methods docstrings."""
+    class_methods: dict[str, str] = {}
+    class_tree = ast.parse(class_and_docstring)
+    for node in class_tree.body:
+        if isinstance(node, ClassDef):
+            for class_node in node.body:
+                if isinstance(class_node, FunctionDef):
+                    class_methods[class_node.name] = ast.get_docstring(class_node)
+    return class_methods
+
+
 def make_docstring_node(docstr: str):
     constant_str: Constant = Constant(docstr)
     return Expr(value=constant_str)
@@ -66,7 +125,6 @@ def get_module_source_code(module_path: str) -> str:
 
 
 def add_module_code_to_queue(module_path: str, module_source_queue: Queue):
-    print(module_path)
     module_src: str = ''
     if module_path:
         module_src = get_module_source_code(module_path)
@@ -119,9 +177,7 @@ def get_all_modules(config: Config, module_source_queue: Queue) -> None:
         directory_iterator: DirectoryIterator = DirectoryIterator(config=config)
         for modules in directory_iterator:
             for module in modules:
-                add_module_to_queue(
-                    module, module_source_queue
-                )
+                add_module_to_queue(module, module_source_queue)
 
 
 def save_processed_file(file_path: str, processed_module_code: str) -> None: